+++ src/index.ts
@@ -28,6 +28,7 @@ export * from './components/animate-slide';
 export * from './components/animate-split';
 export * from './components/animate-typing';
 export * from './components/animate-zoom';
+export * from './components/area-chart';
 export * from './components/aspect-ratio';
 export * from './components/avatar';
 export * from './components/badge';